scan: rename duckdb_native_scan_task -> duckdb_native_decoder (kevkrist)
The file is the split decoder, not a 'scan task' in the old
task-infrastructure sense; the duckdb_native_scan_task name was easy to
confuse with parquet_scan_task and the legacy scan-task path. Rename the
.cpp/.hpp (git mv, history preserved), update the include in the operator
and the self-include, the CMake source entry, and a stale prose comment.
No symbol or behavior changes.
